bitcoin-qt splash screen no longer "not on top"
bitcoin-qt from 0.8.1 to 0.8.3, on windows anyway, always looked quite
broken if any other window happened to partly or fully cover it and was
then with drawn. Could be a pop up from some other program.

I added a one line change for Windows only, but if the Mac or Linux
could benefit, it is obvious. It's just a window initializer default
option. While I was there, the splash screen, now visible seems to be
dead when 'Verifying blocks...". So I animated it a little, but it could
be animated in many different ways.

The idea is to show that it is not frozen, dead, locked, crashed, etc.
but actually doing something with the last two days worth of blocks.
Windows programs don't have a stellar reputation :) Just trying to make
bitcoin-qt more appealing to the masses>

Ron