A new splash screen that has a border.
I added a thin black border to the splash.png file so that when it is
shown above a window that is mostly white, as is the case often in
Windows, it doesn't merge into the background. When displayed on a black
background it looks the same as before, and on any other background it
looks good too.