bitcoind on windows EnvShutdown exception(22) error upon exit fixed
There was no windows code to trap Ctl-C for the bitcoind code. There is
now, and it works. Isee the destructors in CNetCleanup and CMainCleanup
now firing! Maybe this has something to do with a Mac irregular
corruption problem? The Bitcoin-qt "quit" code for 0.8.3 & 0.8.4 still
doesn't call StartShutdown() or something like that in the windows
version. This code is for Windows only. I don't know the state of the
system for Linux, Mac, etc.